input: introduce uniqueDependency policy
Traditionally it was allowed to name a dependency more than once in a
recipe. On the other hand the semantics were not well defined. The
result was picked up only once. Due to the multiple references different
variants of the dependency could be created, though. This was detected
only if the result of the dependencies was used. Otherwise this created
unaddressable packages that cannot be built individually. It is also
possible that, even if the packages themself are of the same variant,
they might provide different dependencies or variables upwards. This is
handled but not easily detectable by the user.

Setting the uniqueDependency policy to the new behavior will halt the
parsing with an error when a dependency is detected multiple times
instead of showing just a warning.

Fixes #193.