input: refactor package calculation
Instead of creating the "real" Package/Step objects during package
calculation and inferring the location independence Core* objects we got
the other way around. Now all package calculation is done on the Core*
objects and the concrete Package/Step objects are only created on demand
as the graph is traversed by other modules.

This breaks the plugin API and string functions a bit because we are not
able anymore to supply real 'Tool' or 'Sandbox' objects. Hopefully this
part of the API is not used by anyone.