jenkins: preserve permissions when extracting artifacts
When building locally the Python tarfile module is used to extract
artifacts. This module will always restore the file mode. When building
on Jenkins GNU tar is invoked that (unless run as root) will not
restore the permissions fully. This leads to inconsistent builds if the
set-user-ID, set-group-ID or sticky permissions are used.

Pass the -p/--preserve option to tar to always set permissions
according to the tar file.