bob: force '0.15-dev-' version string
During development of the next version it should be possible to require
already the new feature or plugin API. But because there is no tag yet
the 'git describe' string will still be the previous version with some
additional commits.

Hard code the upcoming version and just append the current git state. In
case of a release the old logic should be restored.