input: add compat layer for pre-0.15 plugin API
Due to the redesigned package calculation it is not possible to pass the
bob.input.Tool or bob.input.Sandbox objects to the plugins anymore. In
case a plugin requires the old API apply a wrapper that at least that fills
the expected parameters with empty values.

This still changes the behaviour and recipes relying on the removed
parameters will have to be refactored. But at least the parsing should
not crash. A warning is displayed for every usage of a deprecated API.