build: parallel build support
Uses coroutines to schedule more than one job at the same time. Whenever
multiple steps are cooked each step is cooked in its own asyncio task.
The original task then waits for all the subtasks to finish. For every
step that is currently in flight the task is temporarily recorded in a
global map to not falsely spawn multiple tasks for the same step. If the
step succeeds the map entry is removed, otherwise it is kept so that
other tasks do not spawn it again but instead get the exception result.

If build errors happen they are recorded in global list of build errors
that is examined when the main task finshed. Unfortunately it is not
possible to pass the errors as result because the task cancelation will
overwrite any result starting with Python 3.6.

When the user presses Ctrl+C the main task is cancelled which in turn
cancels all subtasks.

When doing a sequential build (-j1) only the main task is ever used and
no further tasks are spawned. This is done to keep the old "depth-first"
behaviour of sequential builds.