archive: fix Python 3.6 compatibility
Starting with Python 3.6 the http.client module does not add a
"Content-Length" header by itself if the body is passed as a file.
Instead it will fall back to chunked encoding which is probably ok for
most servers but breaks our unit tests. Instead we add the correct
Content-Length header ourself which works on all versions so far.