test/status: fix Python 3.7 compatibility
The re.sub() method changed its behavior in Python 3.7 to probably
better match perl. There was a bug report about this change that hits
us here (https://bugs.python.org/issue33585) but it was closed as "not a
bug".

In fact the same difference exists between sed and perl:

$ echo -n foobar | sed -e 's/.*/git/g'
git
$ echo -n foobar | perl -np -e 's/.*/git/g'
gitgit