input: implement 'mergeEnvironment' policy
The 'environment' and 'privateEnvironment' sections of the recipes and
classes it inherits from are merged when the packages are calculated.
Traditionally this was done on a key-by-key basis without variable
substitution. Keys from the recipe or an inherited class would simply
shadow keys from later inherited classes. This had the effect that the
definitions of later inherited classes were lost.

The new behavior is to make all environment keys eligible to variable
substitution. The definitions of the recipe has the highest precedence
(i.e. it is substituted last). Declarations of classes are substituted
in their inheritance order, that is, the last inherited class has the
highest precedence.

Instead of having one resolved __varSelf (resp. __varSelfPrivate) dict
the code now has a list of dicts. Every item in the list is substituted
with the previous environment. If the policy is not set (of set to the
old behavior) the list will have at most one entry where this entry is
created according to the old rules.