status: flag directories that will move to the attic
If the recipe is changed for a SCM checkout it will be moved to the
attic when the checkout is run the next time. In this case the precise
information about the original checkout is lost. Consequently "bob
status" ignored the directory but was silent about doing so. We now flag
the directory as "A" (as in "attic") so that the user knows about the
stale state.

Relates to #220.