state: store SCM properties for each checkout
Except from the information taken from the recipes Bob had no knowledge
about the specific properties of an SCM checkout. The state only held a
checksum that allowed to check if the recipe information still matches.

By storing the SCM properties too in the directory state we will be able
to reconstruct the SCM object even without the recipe. This enables
status checks for existing directories in the future when the recipe is
not available or does not match anymore.