scm: return SCM status as object
This commit establishes a common definition of the SCM status as
ScmStatus object. At its heart the previous single letter 'shortStatus'
flags are stored as set of ScmTaint enum values. They are used
consistently by all SCM implementations. The decision if the SCM is in a
clean/dirty/error state is then just a matter of the set flags. It helps
consistency between the different SCMs and removes code duplication. The
old 'longStatus' is gathered as description field.

The ScmStatus objects are made composable. Separate code parts can thus
supplement the status as needed.

The 'empty' status is removed. If the SCM directory does not exist then
the status() method should not be called. Check at the calling sites
instead of each SCM implementation separately.