status: rewrite status command
Make the package argument optional. If no package is given then all
known checkouts are scanned. This will only give meaningful results if
the project stored the SCM properties for the checkout, though.

Also add an --attic switch to display information about the attic
directories too. This also requires that the project stored the relevant
information. The switch is orthogonal to the package argument. If a
package is selected then all attic directories of this package are
displayed. Otherwise all known attic directories are scanned.

The output verbosity is adjusted too. The package path is only printed
if a STATUS line for the package is emitted. Adding a --show-clean
switch allows the user to print all scanned packages without printing
the detailed information (-v) too.

Fixes #220.