generators: shorten default project directory name
By default the package path was used for the project directory. If the
built package was deep in the hierarchy this could lead to excessively
long path names. Instead use the project name (which defaults to the
package name) as directory name too.

This could theoretically create clashes if different variants of a
recipe are built simultaneously. But the user can handle that by the
--destination option rather than creating excessively long paths by
default.

Fixes #222.