builder: make sure error handler fails the build
If the ERR trap handler is called we have to fail the build. The
current error handler relies on the previously defined
'bob_handle_error' function. If the executed user script starts some
other sub-script that inherits the trap handler but *not* the function
then the trap handler will silently drop the original error.

Do an unconditional exit if bob_handle_error returns or is undefined.