build: use start method "fork" on MSYS platforms
instead of "forkserver", to start a new process.
Start method "forkserver" is only "available on Unix platforms, which
support passing file descriptors over Unix pipes" (see documentation of
"multiprocessing" module, section "Contexts and start methods"). This is
not the case for MSYS and leads to a "RuntimeError: received 0 items of
ancdata" in "/usr/lib/python3.7/multiprocessing/reduction.py", line 161,
in recvfds" for Python 3.7.