clean: extend to develop mode and attic directories
Extend the clean command to also purge unused workspace directories from
develop mode builds. This commit also adds a mode that removes attic
directories.

At the same time the default is changed to 'develop' mode. This
streamlines the behaviour with the other commands in Bob that also work
in 'develop' mode by default. Together with the now added --(no-)sandbox
options it should be consistent with the rest.

If the user removes source workspaces (-s or --attic) Bob will now check
the SCMs in these directories for unsaved changes. The directory will
only be deleted if no change is found.

Fixes #101.