status: fix attic selection
The attic directories are located next to the workspace, not inside it.
The code relied on the fact that attic paths were stored as relative
paths from the workspace. This is a bad idea because the attic cannot be
reached anymore if the workspace was deleted. Instead compute a proper
prefix that is compared by removing the last path element which is
always '/workspace'.