input: introduce 'netAccess' tool property
Starting with the introduction of the 'offlineBuild' policy the network
access is usually restricted during build and package steps. There
might be tools that need network access and are used during these
steps, though. This can be configured with the {build,package}NetAccess
recipe properties but if the recipe does not know the exact tool it
might not want to set these properties unconditionally.

The main use case are proprietary compilers that need to talk to a
license server. The recipe should not bother which compiler is used
exactly.