input: implement tool environment definitions
Certain tools (e.g. the C-compiler) and their associated environment
variables (CC, LD, ...) are related to each other and must be defined
and used together. Previously they were provided more or less
independently of each other by a tools recipe. This has several
drawbacks:

 * Not all recipes that define the same tool (e.g. the compiler) are
   providing the same set of environment variables.
 * Different tools might provide the same variables.

By defining the environment variables with the tool their value can be
picked up where it is actually used. This does not pollute the
environment with variables that are only used in conjunction with the
tool and are not used elsewhere.

To prevent ambiguities each variable must only be defined by one tool.
The parsing will fail if two tools are used in the same recipe that
define the same variable. The reason is that Bob assumes that tools are
independent of each other and no particular order is defined among them.