input: fix sandboxInvariant policy
Setting the sandboxInvariant policy to the new behaviour had exactly the
opposite effect as intended. It caused artifacts of sandbox and
non-sandbox builds to be always incompatible to each other. Fix that by
always leaving out the sandbox in the variant-id and build-id
calculation if the project is on the new behaviour.