build: add fingerprint support
The fingerprint feature is used to track all dependencies of the recipes
to the host environment in a generic manner. This may start from the
host C-compiler and reach up to some libraries that are used from the
host. Because such external dependencies limit the exchange of artifacts
between hosts the fingerprint is attached to binary artifacts to prevent
unintended sharing between incompatible systems.

The fingerprints are also used to detect relevant changes of the host
environment. A package is always rebuilt if the fingerprint changes.
The fingerprint is thus like an external build trigger. There are two major
cases why we want to rebuild when the fingerprint changes:

 * The package was built inside the sandbox and now the user builds
   outside of the sandbox (or vice versa).
 * The package was built without sandbox and the fingerprint changes. We
   probably want to rebuild because the external dependency has changes
   its state.

If a package is not relocatable the workspace path is also some kind of
host dependency. We piggy-pack the mechanism and mix-in the workspace
path into the fingerprint. This enables safe exchange of such artifacts
that were previously only up-/downloaded when building inside a sandbox.