scm: streamline ScmStatus description handling
The detailed description of a ScmStatus is now stored internally
per-flag instead of one large string. This allows to selectively show
descriptions of some flags instead of all.

The verbosity handling of status lines is handled per flag. Each status
flag has an associated severity. The detailed description of a flag is
shown on the next verbosity level. There are two special tweaks: with
'--show-overrides' the 'overridden' flag severity is selectively
increased to NORMAL; with '--show-clean' the headline is always shown
but the detailed description severities are not changed.