path/archive: fix consecutive infix expressions
Apparently pyparsing will call the handler of a binary infix operation
with all operands instead of only two when the same operation is used
more than once at the same rank. Example:

  term1 || term2 || term3

In this case the handler is called with:

  [[term1, '||', term2, '||', term3]]

Instead of fixing all existing parsers provide a wrapper that does
recursively call the handler with only two terms at once.