input: compute exec path by referrer
The execution path of a step depends on whether a sandbox is used or
not. That criteria has to be based on the state of the actually executed
step. In case the execution path of the dependencies of a step are computed
this decision must be based on the very same step.

The old logic looked only on the state of the step object and not the
one that is referring to it, tough. This is wrong if the dependency is
built in a sandbox and the dependent package is built outside. The
dependent package saw paths that were valid only in the sandbox.

Fixes #237.