git: prevent cloning orphaned tags
Git does not normally fetch tags that are not on a remote branch. The
only exception is the initial clone that fetches all remote tags. This
is inconsistent and we cannot be sure if such orphan tag is on any
remote or was just created locally.

If Bob has to check out a certain commit or tag the fetch refspec is
amended to also explicitly include this object besides all branches.
When cloning a branch we now do the fetch and checkout manually to get
only tags from branches.