git: ignore checked out tag as unpushed local commit
If a tag is not on a remote branch it will be normally seen as
"unpushed". Usually such tags are not cloned at all but if the user
configured such tag in the recipe it would still be cloned. We can
thus safely ignore such a tag on the "unpushed" log calculation.