build: fix checkout-only parallel build
When building with checkout-only the packages are normally not built
and packaged. But if a checkout step uses a tool then this tool must
still be built. This can lead to conflicting jobs for the same package
where on one path it should be skipped and another package path
requires the tool to be built.

Adjust the task logic to treat such jobs differently and, if a
conflicting job is already running, wait for it when starting a new
job.