input: improve fingerprintIf keyword semantic
Instead of implicitly enabling fingerprinting when a 'fingerprint'
script is defined the user now has to enable it explicitly with at least
one 'fingerprintIf'. Bob will also ignore any associated
'fingerprintScript' if the corresponding 'fingerprintIf' evaluates to
false. This prevents execution of fingerprint scripts that should not
be run according to their condition.

If fingerprintIf is unset or null the fingerprinting will not be
enabled for a recipe but the fingerprintScript is retained if any other
part of the recipe enables it. This provides composability where
classes can provide fingerprint script helpers without forcing the
fingerprinting of the package yet.

Also introduce fingperintIf attribute to provideTools. Because the
existence of a fingerprintScript does not enable fingerprinting anymore
a tool must have the ability to actually force fingerprinting.