bob: fix fingerprint behavior
The fingerprint is now fully integrated into the variant- and build-id.
There is no need anymore for a separate fingerprint handing at the
artifact archive and audit side. This also simplifies the build logic a
bit. In case of sandbox builds the variant-id get's a fingerprint so
that packages that are built in different sandboxes are getting separate
workspaces.

Fingerprints are implemented by amending the existing variant- and
build-id. If a fingerprint exists it is appended to the existing 20
bytes SHA1 digest. Both parts can be distinguished easily by slicing the
final digest. This retains the types of the variant- and build-ids while
still enables separate calculations on both fields.

Fingerprints are only transitive for the result of a recipe. Provided
tools or sandboxes do not pass the fingerprint to the users. This is
based on the assumption that tools/sandboxes provide *behaviour* that is
used by the consumers but not actual content.

The actual fingerprinting is made independent of the allRelocatable
policy. This policy is only queried for non-relocatable package steps in
which case the fingerprint is amended by the execution path.

Fixes #239.