input: reject defining variables starting with 'BOB_'
Variable names starting with 'BOB_' have historically been used by Bob
all over the place. Lately some more variables have been added. Instead
of hoping that this does not break existing recipes we prevent the user
from defining any variables with this prefix once and for all.

While this may break existing recipes it is better to do it now instead
of accidentally colliding with user recipes in the future.