build: properly handle --keep-going on -j1
The build scheduling logic is different for sequential builds. It should
be deterministic and strictly work depth first. Otherwise the builder
would jump between different packages, polluting the log and confusing
the user.

Now to support --keep-going in this case we must make sure that we still
descend to all neighbouring dependencies even if one fails. This is
accomplished by scheduling tasks sequentially, wait for them immediately
and gathering the results as the last step.

Fixes #249.