bob: add setuptools installation method
Make Bob a proper member of the Python family. The recommended
installation procedure should be via 'pip'. To keep the possibility to
hack on Bob the bob.develop package is introduced which is *not*
installed. If this package is present Bob assumes that he is used
in-source and automatically builds the bob-namespace-sandbox helper and
manpages.