generators: add Visual Studio 2019 generator
This generator works currently only on MSYS2. Bob and the build is run
on MSYS2 while Visual Studio is running natively. The drawback is that
debugging of applications does not work because they would need to be
built with the MS compiler. Error messages are also not understood by VS
because the paths emitted by the compier refer to MSYS names and not to
native Windows paths.

Fixes #248.