builder: handle tasks for different sandboxes concurrently
If the sandboxInvariant policy is set to the new behaviour it can
happen that the same workspace is possibly used with different
sandboxes (unless a fingerprint is defined for the package). This get's
a real problem if the package is a dependency of a sandbox that is used
to build the package. In this case the build logic got stuck in a
deadlock.

To prevent this the tasks are keyed additionally with the sandbox
variant id. While this prevents a deadlock it might cause the same step
to be executed concurrently in the same workspace. To prevent this the
workspace is locked while the step is executed. Once the lock is
obtained it is re-checked if the step was not already executed in the
meantime.

Fixes #266.