archive: store live build-ids separately for Windows
When running on Windows the result of extracting an archive is not
necessarily the same as on a POSIX system. Specifically the mode bits
are not really supported. This breaks live build-id support in case both
systems are mixed in one binary repository. So let's be honest and treat
both systems differently.