jenkins: prevent permanent HTTP 403 error
Starting with Jenkins 2.186 / 2.176.2 (LTS) any request to a standard
Jenkins installation will result in a

  HTTP error: 403 No valid crumb was included in the request

error. Disabling the CSRF protection in the Jenkins security settings
fixes this error but this is strongly discouraged in public
installations. Another possible workaround is to set the system
property hudson.security.csrf.DefaultCrumbIssuer.EXCLUDE_SESSION_ID to
true. See https://jenkins.io/doc/upgrade-guide/2.176/#SECURITY-626 for
more details.

The reason is the fix of SECURITY-626 / CVE-2019-10353. The CSRF token
is now bound to the session. Previously the token was only checked for
user authentication and IP address. By switching to urllib for the
internal processing the session is now kept in all requests and the
token successfully authenticates on Jenkins side.