builder: do not cancel buildId calculation
If a job fails the overall execution is stopped and no more tasks are
started. Since buildId calculation is also a executed as a task the
buildId of a parallel running successfull job was not calculated and the
build was executed in the next bob build.

Handle this by adding a ignoreExecutionStop argument to __yieldJobWhile.
This argument should be set to true for all 'cleanup' tasks.

Fixes #271.