status: raise severity of A and N flags
The handling of the A and N flags was inconsistent with respect to the S
flag. All these flags are shown by default now. They are all
inconsistencies with respect to the recipes. They should be treated with
equal severity like direct source modifications because these are
consequences of modifications in the recipes or manual changes of the
checkouts.

Fixes #282.