stringparser: support bare variable names
Usually the name of the variable must be enclosed in curly braces when
dereferencing it. In regular POSIX shells this is not needed if the
variable name consists of only letters, numbers and underscores *and* if
the next character is something different to terminate the variable
name. In the light of more compatibility we support this notion too now.