archive: implement sandbox fingerprint caching
With the upcoming sandboxFingerprints policy the fingerprint scripts are
executed in the sandbox too. To reduce the need to build/fetch the
sandbox image just to execute the fingerprint script a *.fprnt file is
introduced to cache the result. Given the same sandbox image and the
same fingerprint script the result will be cached by these files.

Add the methods for local up-/download and for Jenkins uploads.