bob: introduce sandboxFingerprints policy
Bob 0.15 used a shortcut (by using the sandbox variant-id) and did not
execute fingerprint scripts in the sandbox. This saved a bit of
complexity and also relieves us from the need to have the sandbox around
just to execute the fingerprint script.

While the old approach was not producing wrong results it was overly
pessimistic. It prevents sharing of any fingerprinted artifacts between
sandbox and non-sandbox builds even if the fingerprint would be the
same.

It turned out this is undesired and the sandboxFingerprints policy
rectifies this. When set to the new behaviour the fingerprint scripts
will be executed in the sandbox too. A caching of these results by the
artifact cache is also implemented to reduce the need of fetching the
sandbox image. Previously built artifacts will still be found if the
policy is set to the old behaviour.

Fixes #284.