Revert "scm: fix variable substitution in overrides "replace" section"
Starting with 5248fde the regex pattern is also first mangled through
the Bob string substitution. Unfortunately the regex and string
substitution share some meta-characters, namely $ and \. So this created
another layer of indirection and regex patterns that used to work before
are failing. It's also not very user friendly because the typically
incomprehensible regex must be further escaped.

This reverts commits 5248fde69541541aa4f790351221afe823933282,
7d0b3574f830d586a01c856cf4d9e6893906c4a5 (partially) and
99d34b1c29c5591bf8beeafb2e57d6f47af76f53.

Fixes #256.