setup: defer import of sphinx build command
We already name the 'sphinx' package as a build dependency but this
doesn't gain anything if we try to import this package from setup.py
directly. Instead pass a proxy object to the setuptools machinery and
import the sphinxy BuildDoc command only when needed. This should happen
late enough when the build dependency was fetched.

Let's hope setuptools never start to look at the passed command class
directly.