build: run fingerprint script in workspace
In some setups the temporary directory (usually /tmp) is be mounted as
"noexec". This prevents the execution of compiled test applications and
will fail the included "bob-libc-version" and "bob-libstdc++-version"
helpers. We now use the current workspace as base directory for
temporaries because it can safely be assumed that its writable and
executable.