builder: specialize task creation
Basically there are currently two types of tasks: generic ones that need
some sort of wrapping and cooking-tasks that build something for a
particular step and that must not be run twice. Create two dedicated
methods to create the different types instead of overloading it in one
method that behaves very different depending on the arguments.