input: add fingerprintVars policy
When then fingerprintScript mechanism was introduced there was no
dedicated environment handling implemented for them. The simple policy
was to pass all environment variables of the affected step to the
fingerprintScript. Unfortunately this results of the repeated execution
of these scripts if the variables change between packages, even if they
are not used by the fingerprintScript.

The fingerprintVars policy adds the support for the new fingerprintVars
key in the recipes. This key specifies a list of variables that the
fingerprintScript uses. If the fingerprintVars policy is set to the new
behaviour then the environment is filtered and only the variables named
in fingerprintVars are passed to the fingerprintScript. If the policy is
kept on the old behaviour then the fingerprintVars keys are ignored and
all environment variables of a step are passed to the fingerprintScript.