audit: add method to pass extra data from Scm to ScmAudit
Previously the ScmAudit did just work on the information that was
present in the workspace. If the Scm wanted to forward more information
to the Audit then there was no way to do that. Starting with this
commit the Scm could pass extra information and use that when creating
the audit record.

Currently the information is empty.