version: convert git version to PEP 440 compliant format
We used to make git tags in a SemVer compliant manner. We keep that
schema for tags but convert them to a PEP 440 compliant format. The
rationale is that setuptools require the PEP 440 format and it will be
the format that the user sees when looking at "pip". So even if we run
directly from git we should output a compatible version number.