builder: do not ignore cancelled error
If an asyncio task is cancelled it will receive a
concurrent.futures.CancelledError exception. We used to ignore this in
__taskWrapper() but this will silently let the dependent tasks proceed
even though the task was interrupted. This in turn might lead to
exceptions, e.g. during the build-id calculation, because the result
that should have been created by the task is not there.

Make sure that we gracefully stop the affected task and that every
dependent task is then cancelled too.