query-path: Show message on empty output
In case that one or more paths for the given query could not be found,
present a message to the user stating exactly that. Inform the user
when a query matched no packages at all, too. Bob's silence in these
cases could be quite confusing.

Note that since query-path is mainly used for scripting purposes and
these scripts may rely on the empty output, the message is printed to
stderr.

Update the copyright statement whilst we're at it.