input: fix fingerprinting of build steps
The build step may have fewer tools than the package step. This must be
taken into account when merging the fingerprint scripts of the tools.
Even worse, if the fingerprinting is enabled by a tool that is *only*
used in the package step the build step should not even be fingerprinted
at all. Previously the build step would always be fingerprinted if the
package step has a fingerprint. Also the wrong fingerprint scripts may
have been used in the build step if the number of tools differ compared
to the package step.

This may break some binary artifacts.