bob: do not generate bash scripts for SCMs / checkoutAsserts
Move everything related to bash into a separate language module. The
rationale is to keep the core of Bob independent from the actually
executed scripting language. The languages module provides an interface
the get the language object that is used for script invocation.
Currently only bash is supported but the languages module can easily
extended with additional languages.

All building is done through an Invoker object that takes care of the
dirty details to call external executables. This object will also
handle the output redirection, logfile creation, exit code checking and
directory handling. If executing a step manually the invoker is called
explicitly via an internal command.

The SCMs do not generate bash scripts anymore. The are called before
the main script and use the invoker to execute the required external
commands. The same is true for the checkoutAssert stuff that is run
after the checkout.