input: add multi language support
The recipes may use multiple languages. The default language is 'bash'
but this may be overridden by the 'scriptLanguage' setting in
config.yaml. The default language is used to call the right interpreter
for the checkoutScript/buildScript/packageScript/fingerprintScript
scripts.

For each *Script entry there is a variant for the different supported
languages. For example the buildScript has the buildScriptBash and
buildScriptPwsh siblings. They hold the scripts for the respective
languages. This schema extends to the other scripts too.

The selection which language is used at build time is done during
execution. By default bash scripts are used. The language may be
configured globally in config.yaml by setting 'scriptLanguage' or on in
the recipe/class by the same key.