test: fix $PATH in sandbox-fingerprints test
Having a sandbox without any $PATH is not supported. At least the
language interpreter must be in $PATH. Otherwise the step cannot be
executed. This has never been observed yet because /bin/bash was hard
coded as absolute path but with the advent of additional scripting
languages this does not hold true anymore.