tty: re-initialize win32 tty
The parsing of VT100 codes must be enabled explicitly on the console in
Windows 10. Unfortunately the MSYS2 executables will unconditionally
disable the support upon termination. This will break our color output.
Hence we restore the feature flags every time a spawned process exits.

This is not a problem on parallel builds. In this case the
stdin/-out/-err file descriptors are pipes and the MSYS2 code won't
mess with the console.