input: fix fingerprinting of build steps (again)
If the build step does not use a tool that is used in the package step
then things could still go wrong. If the same build step is reachable
through different path where the not-used tool is only available on one
path then the bit mask will be wrong on the other path.

The fix is to always iterate the list of consumed tools of the package
step. This will lead to the same bit offsets even if the tool is
available or not.