languages: dump PowerShell environment as JSON
We now dump the environment as well as all internal variables into a
JSON string. This also retains the structure of the internal variables
(they are objects by themselves) and makes everything nicely machine
readable. The overall structure looks like the following:

  {
    "Env" : {
      "NAME" : "VALUE"
    },
    "Vars" : {
      "name" : ...
    }
}

The values of environment variables are always strings. The values of
(internal) variables can be of any type.

Fixes #301.