jenkins: store shared packages under their build-id
On Jenkins artifacts that are produced by recipes that have a "shared:
True" are shared globally between all jobs on a slave. The artifact was
previously indexed by variant-id. The variant-id is stable since a long
time but the build-id calculation changed recently, e.g. by the
sandboxInvariant policy. Now these artifact are used from the shared
location even though their build-id is potentially different. This will
create artifacts that are not found by builds outside of Jenkins and
that have a wrong build-id.

We now index shared packages by their build-id. This will make sure that
always the right artifact is reused. Because the indexing scheme changed
we bump the generation just in case there is some aliasing.

Fixes #287.