archive: separate all artifacts of the major platforms
There are three major platforms that all behave subtly differently:

 * POSIX systems such as Linux
 * Native Windows
 * MSYS2 on Windows

Given the same archive that is extracted the hash sum of the result may
be different. This is mostly caused by the different supported file
modes.

Replaces and amends c6a0c2b6 which only covered the prediction of
build-ids.

Fixes #304.