archive: distinguish symlink capability on Windows
Some scripts and archive utilities will behave differently if it is
possible to create symlinks on Windows or not. This capability is not
granted by default but only available to administrative shells or user
accounts that were given the SeCreateSymbolicLinkPrivilege privilege.
Threat builds with and without the symlink capability differently
because there there would be false sharing if symlinks are actually
used.

Fixes #303.