jenkins: set BOB_HOST_PLATFORM according to Jenkins config
The Jenkins (slaves) that build the project do not need to have the same
platform architecture than the one that is configuring. Set the
BOB_HOST_PLATFORM identification either to Linux or MSYS2 depending on
the --windows switch. Native Windows builds are not working yet so they
are currently out of scope.

Fixes #307.