builder: only create new checkout audit if something changed
Merely executing the checkout step does not require the audit trail to
be re-created unless something changed. The usual case is that the SCMs are
checked for updates. If no updates were found and the checkout workspace
did not change the old audit trail record is still valid. This saves
significant time for large checkouts because no SCM status need to be
scanned.