bob: store executed script next to workspace
Since the multi-language refactoring the script file that is executed by
the interpreter was put in a temporary directory and removed after the
invocation. This is sub-optimal when debugging hard script failures.

We now pass a hint to the invoker where the script file should be
stored. The scripting language is free to follow it or add a proper file
ending as PowerShell needs. The actual chosen location is passed back to
the invoker to setup the sandbox accordingly (is used).

Fixes #320.