build: add option to build without audit trail
The generation of the audit trail is usually barely noticeable. But if
a large number of repositories is checked out it can add a significant
overhead nonetheless. Add a -A / --without-audit option so that the
user can skip the generation of an audit trail to save this time.

Without an audit trail it is not possible anymore to upload an artifact
because vital information is missing. Consequently deny the generation
of an audit trail if the audit trail of a dependency is missing or if
it cannot be read. Otherwise the information would be incomplete.
Historically this was accepted but only because the rest of the logic
could not cope with a missing audit trail at all.