input: relax scmOverrides restrictions
The values of checkoutSCM keys are not only strings. They can be
booleans (e.g. sslVerify) and integers too (shallow). Allow any object
in the scmOverrides block and let the actual validation be done by the
schema checks.

Likewise allow to delete any key. Previously a partial list of supported
keys could be deleted but other valid keys could not.