bob: guard main entry against module loading
Starting with 226b9356 the audit engine is also running an asyncio event
loop. When running on POSIX systems this implies to have multiple
processes (due to the usage of the multiprocessing module) that import
the main entry point module. We must make sure to not execute code if
the bob-hash-engine "module" is imported instead of being executed.

Fixes #328.