input: add tool remapping support for dependencies
For each dependency in the "depends" list an existing tool can be
remapped to a new name, possibly replacing an existing tool. This can
be used to replace the target toolchain with the current cross
toolchain for a dependency.

The change requires that the CoreRef class needs to cope with the
functionality too. The user might remap a tool that was defined outside
of the current recipe. This is encoded as a string in the diffTools.