bob: handle SIGINT in executor context
The multiprocessing module forces us to fiddle with the signal
handlers. When the user presses ctrl+c we do not want to see random
backtraces from this module so we block SIGINT by default. When
executing code we have to restore the standard Python signal handler,
though. Otherwise KeyboardInterrupt will not be raised and the
exception handling does not work. On the other hand we must not keep
the standard handler or else we will see the backtraces for idle
workers again.

Fixes #345. The code to clean stale temporary files was already there.
It just wasn't executed on ctrl+c because SIG_DFL will instantly kill
the process and not invoke the exception handling.